+++ changelog.md
@@ -1,3 +1,4 @@
 # Changelog
 
-Updated summary
+**./src/output-sql-files.js**
+* Called 'saveSql' function for all tables.